requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
redis==5.0.8
supabase==2.7.4
groq==0.11.0
chromadb==0.6.3
langchain==0.3.1
langchain-groq==0.2.0
langchain-community==0.3.1
sentence-transformers==3.1.1
python-multipart==0.0.9
aiofiles==24.1.0
beautifulsoup4==4.12.3
requests==2.32.3
apscheduler==3.10.4
